Moved NWStatus to core/status.py
